+++ web/vite.config.ts
@@ -1,14 +1,13 @@
-import tailwindcss from '@tailwindcss/vite';
-import { sveltekit } from '@sveltejs/kit/vite';
-import { defineConfig } from 'vite';
-import wasm from 'vite-plugin-wasm';
-import top_level_await from 'vite-plugin-top-level-await';
+import tailwindcss from "@tailwindcss/vite";
+import { sveltekit } from "@sveltejs/kit/vite";
+import { defineConfig } from "vite";
+import wasm from "vite-plugin-wasm";
 
 export default defineConfig({
-	plugins: [tailwindcss(), sveltekit(), wasm(), top_level_await()],
+	plugins: [tailwindcss(), sveltekit(), wasm()],
 	server: {
 		fs: {
-			allow: ['../pkg']
-		}
-	}
+			allow: ["../pkg"],
+		},
+	},
 });