+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## Next
+- Update `fastnum` dependency
+- Fix conversion of `torr`, `btu/s` and `btu/h`
+- Fix unit spellings
+- Fix spelling of `decares`
+
 ## 3.0.0 - 2025 May 30
 - Create a [website interface](https://cpc.kasper.space)
 - Add wasm version